Refactor GCS reporter: rename to gcs_artifact, update config key to gcs_artifacts, and apply readability/design improvements
