+++ evalbench/reporting/__init__.py
@@ -1,7 +1,7 @@
 from .csv import CsvReporter
 from .bqstore import BigQueryReporter
 from .report import Reporter
-from .gcs import GcsReporter
+from .gcs_artifact import GcsReporter
 
 
 def get_reporters(reporting_config, job_id, run_time) -> list[Reporter]:
@@ -15,7 +15,7 @@ def get_reporters(reporting_config, job_id, run_time) -> list[Reporter]:
     if "csv" in reporting_config:
         reporters.append(CsvReporter(
             reporting_config["csv"], job_id, run_time))
-    if "gcs" in reporting_config:
+    if "gcs_artifacts" in reporting_config:
         reporters.append(GcsReporter(
-            reporting_config["gcs"], job_id, run_time))
+            reporting_config["gcs_artifacts"], job_id, run_time))
     return reporters
